Added pylesa regression test
