+++ pylesa/__main__.py
@@ -1,6 +1,13 @@
+import logging
 import typer
 
 from .main import main
 
+LOG = logging.getLogger(__name__)
+
 if __name__ == "__main__":
-    typer.run(main)
+    try:
+        typer.run(main)
+    except Exception as e:
+        LOG.error(e, exc_info=True)
+        raise e
