+++ tests/data/fixed_order/outputs/KPIs/KPI_economic_fixed_order.csv
@@ -1,0 +1,5 @@
+hp_sizes,ts_sizes,capital_cost,opex,cost_of_heat,cost_elec,lifetime_cost,levelised_cost_of_heat,levelised_cost_of_energy
+2000.0,200.0,1305.2115231339765,295.1238626624039,0.0358182508845421,295.12386266240384,7207.688776382053,0.05905110529198202,0.07345915360992547
+2000.0,400.0,1348.6884642170814,290.105501777348,0.03397989860969277,290.10550177734797,7150.798499764041,0.058189109501953974,0.07287934062151126
+3000.0,200.0,1905.2115231339765,295.5167625846827,0.03559198833076488,295.51676258468274,7815.546774827631,0.06826130898132525,0.07965430651763042
+3000.0,400.0,1948.6884642170814,288.22870558323734,0.0328144666510219,288.2287055832374,7713.262575881829,0.06671150901340316,0.07861184881512133
