+++ tests/data/fixed_order/outputs/KPIs/KPI_technical_fixed_order.csv
@@ -1,0 +1,5 @@
+hp_sizes,ts_sizes,local_RES_used,total_RES_self_consumption,grid_RES_used,total_RES_used,heat_met_RES,demand_met_RES,HP_met_RES,HP_size_ratio,HP_utilisation,days_storage_content
+2000.0,200.0,88.30294893623967,323234.1366034354,0.0,323234.1366034354,10.582527014356963,10.425086253330651,10.582527014356963,1.97,110.00000000000001,0.513716756390289
+2000.0,400.0,90.3254987523351,330637.70750812953,0.0,330637.70750812953,11.037688761721205,10.595749421333768,11.037688761721205,1.99,111.00000000000001,1.027433512780578
+3000.0,200.0,87.9314819481215,321874.37667890836,0.0,321874.37667890836,10.375384671248005,10.333863173601465,10.375384671248005,2.75,111.00000000000001,0.513716756390289
+3000.0,400.0,89.58210346864313,327916.49903690687,0.0,327916.49903690687,10.611430924314549,10.405693956540707,10.611430924314549,2.95,112.99999999999999,1.027433512780578
