+++ tests/data/fixed_order/outputs/KPIs/output_fixed_order.csv
@@ -1,0 +1,5 @@
+hp_sizes,ts_sizes,sum_hp_output,sum_hp_usage,scop,sum_aux_output,sum_ed_import,sum_RES,sum_import,sum_export
+2000.0,200.0,3618354.241726054,1494541.7329334118,2.42104597147925,-8.992806499463768e-14,1416330.5665570789,366051.3499235806,2777307.468029977,42817.21332014522
+2000.0,400.0,3666425.2758746375,1514475.0882156675,2.420921482567512,-5.440092820663267e-14,1416330.5665570789,366051.3499235806,2789837.2524075382,35413.642415451046
+3000.0,200.0,3652395.554119806,1508753.7424291053,2.4208029789138554,-1.1834977442504169e-13,1416330.5665570789,366051.3499235806,2792879.2374501973,44176.973244672234
+3000.0,400.0,3739447.123889275,1545318.0548741154,2.419855972105301,-1.1124434706744069e-13,1416330.5665570789,366051.3499235806,2823401.4275372084,38134.85088667371
